+++ doc/source/train/examples/pytorch/torch_fashion_mnist_example.rst
@@ -1,0 +1,6 @@
+:orphan:
+
+torch_fashion_mnist_example
+===========================
+
+.. literalinclude:: /../../python/ray/train/examples/pytorch/torch_fashion_mnist_example.py
