+++ doc/source/train/examples/pytorch/torch_regression_example.rst
@@ -1,0 +1,6 @@
+:orphan:
+
+torch_regression_example
+========================
+
+.. literalinclude:: /../../python/ray/train/examples/pytorch/torch_regression_example.py
