+++ doc/source/train/examples/pytorch/tune_cifar_torch_pbt_example.rst
@@ -1,0 +1,6 @@
+:orphan:
+
+tune_cifar_torch_pbt_example
+============================
+
+.. literalinclude:: /../../python/ray/train/examples/pytorch/tune_cifar_torch_pbt_example.py
