+++ doc/source/train/examples/pytorch/tune_torch_regression_example.rst
@@ -1,0 +1,6 @@
+:orphan:
+
+tune_torch_regression_example
+=============================
+
+.. literalinclude:: /../../python/ray/train/examples/pytorch/tune_torch_regression_example.py
