--- doc/source/train/examples/tensorflow_mnist_example.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-tensorflow_mnist_example
-========================
-
-.. literalinclude:: /../../python/ray/train/examples/tensorflow_mnist_example.py
