--- doc/source/train/examples/tensorflow_regression_example.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-tensorflow_regression_example
-=============================
-
-.. literalinclude:: /../../python/ray/air/examples/tf/tensorflow_regression_example.py
