+++ doc/source/train/examples/tf/tensorflow_mnist_example.rst
@@ -1,0 +1,6 @@
+:orphan:
+
+tensorflow_mnist_example
+========================
+
+.. literalinclude:: /../../python/ray/train/examples/tf/tensorflow_mnist_example.py
