+++ doc/source/train/examples/tf/tensorflow_regression_example.rst
@@ -1,0 +1,6 @@
+:orphan:
+
+tensorflow_regression_example
+=============================
+
+.. literalinclude:: /../../python/ray/train/examples/tf/tensorflow_regression_example.py
