+++ doc/source/train/examples/tf/tune_tensorflow_mnist_example.rst
@@ -1,0 +1,6 @@
+:orphan:
+
+tune_tensorflow_mnist_example
+=============================
+
+.. literalinclude:: /../../python/ray/train/examples/tf/tune_tensorflow_mnist_example.py
