--- doc/source/train/examples/torch_fashion_mnist_example.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-torch_fashion_mnist_example
-===========================
-
-.. literalinclude:: /../../python/ray/train/examples/torch_fashion_mnist_example.py
