--- doc/source/train/examples/torch_regression_example.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-torch_regression_example
-========================
-
-.. literalinclude:: /../../python/ray/air/examples/pytorch/torch_regression_example.py
