--- doc/source/train/examples/tune_cifar_torch_pbt_example.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-tune_cifar_torch_pbt_example
-============================
-
-.. literalinclude:: /../../python/ray/train/examples/tune_cifar_torch_pbt_example.py
