--- doc/source/train/examples/tune_tensorflow_mnist_example.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-tune_tensorflow_mnist_example
-=============================
-
-.. literalinclude:: /../../python/ray/train/examples/tune_tensorflow_mnist_example.py
