--- doc/source/train/examples/tune_torch_regression_example.rst
@@ -1,6 +1,0 @@
-:orphan:
-
-tune_torch_regression_example
-=============================
-
-.. literalinclude:: /../../python/ray/air/examples/pytorch/tune_torch_regression_example.py
