+++ python/ray/air/BUILD
@@ -20,143 +20,6 @@ py_test(
     deps = [":ml_lib"]
 )
 
-
-py_test(
-    name = "distributed_sage_example",
-    size = "small",
-    main = "examples/pytorch_geometric/distributed_sage_example.py",
-    srcs = ["examples/pytorch_geometric/distributed_sage_example.py"],
-    tags = ["team:ml", "exclusive", "gpu"],
-    deps = [":ml_lib"],
-    args = ["--use-gpu", "--num-workers=2", "--epochs=1", "--dataset=fake"]
-)
-
-py_test(
-    name = "horovod_cifar_pbt_example",
-    size = "small",
-    srcs = ["examples/horovod/horovod_cifar_pbt_example.py"],
-    tags = ["team:ml", "exlusive"],
-    deps = [":ml_lib"],
-    args = ["--smoke-test"]
-)
-
-py_test(
-    name = "horovod_pytorch_example",
-    size = "small",
-    srcs = ["examples/horovod/horovod_pytorch_example.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"],
-    args = ["--num-epochs=1"]
-)
-
-py_test(
-    name = "horovod_tune_example",
-    size = "small",
-    srcs = ["examples/horovod/horovod_tune_example.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"],
-    args = ["--smoke-test"]
-)
-
-py_test (
-    name = "huggingface_basic_language_modeling_example",
-    size = "medium",
-    srcs = ["examples/huggingface/huggingface_basic_language_modeling_example.py"],
-    args = ["--smoke-test", "--num-epochs 3"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"]
-)
-
-py_test(
-    name = "tensorflow_regression_example",
-    size = "medium",
-    main = "examples/tf/tensorflow_regression_example.py",
-    srcs = ["examples/tf/tensorflow_regression_example.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"],
-    args = ["--smoke-test"]
-)
-
-py_test(
-    name = "tensorflow_mnist_example",
-    size = "medium",
-    main = "examples/tf/tensorflow_mnist_example.py",
-    srcs = ["examples/tf/tensorflow_mnist_example.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"],
-    args = ["--smoke-test"]
-)
-
-# py_test(
-#     name = "tensorflow_autoencoder_example", # REGRESSION
-#     size = "medium",
-#     main = "examples/tf/tensorflow_autoencoder_example.py",
-#     srcs = ["examples/tf/tensorflow_autoencoder_example.py"],
-#     tags = ["team:ml", "exclusive"],
-#     deps = [":ml_lib"],
-#     args = ["--smoke-test"]
-# )
-
-py_test(
-    name = "torch_fashion_mnist_example",
-    size = "medium",
-    main = "examples/pytorch/torch_fashion_mnist_example.py",
-    srcs = ["examples/pytorch/torch_fashion_mnist_example.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"],
-    args = ["--smoke-test"]
-)
-
-py_test(
-    name = "torch_fashion_mnist_example_gpu",
-    size = "medium",
-    main = "examples/pytorch/torch_fashion_mnist_example.py",
-    srcs = ["examples/pytorch/torch_fashion_mnist_example.py"],
-    tags = ["team:ml", "exclusive", "gpu"],
-    deps = [":ml_lib"],
-    args = ["--use-gpu"]
-)
-
-py_test(
-    name = "torch_regression_example",
-    size = "medium",
-    main = "examples/pytorch/torch_regression_example.py",
-    srcs = ["examples/pytorch/torch_regression_example.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"],
-    args = ["--smoke-test"]
-)
-
-py_test(
-    name = "torch_linear_example",
-    size = "small",
-    main = "examples/pytorch/torch_linear_example.py",
-    srcs = ["examples/pytorch/torch_linear_example.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"],
-    args = ["--smoke-test"]
-)
-
-py_test(
-    name = "tune_tensorflow_mnist_example",
-    size = "medium",
-    main = "examples/tf/tune_tensorflow_mnist_example.py",
-    srcs = ["examples/tf/tune_tensorflow_mnist_example.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"],
-    args = ["--smoke-test"]
-)
-
-py_test(
-    name = "tune_torch_regression_example.py",
-    size = "small",
-    main = "examples/pytorch/tune_torch_regression_example.py",
-    srcs = ["examples/pytorch/tune_torch_regression_example.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"],
-    args = ["--smoke-test"]
-)
-
 # --------------------------------------------------------------------
 # Tests from the python/ray/air/tests directory.
 # Covers all tests starting with `test_`.
