+++ python/ray/train/BUILD
@@ -29,8 +29,8 @@ py_test(
 py_test(
     name = "tensorflow_quick_start",
     size = "medium",
-    main = "examples/tensorflow_quick_start.py",
-    srcs = ["examples/tensorflow_quick_start.py"],
+    main = "examples/tf/tensorflow_quick_start.py",
+    srcs = ["examples/tf/tensorflow_quick_start.py"],
     tags = ["team:ml", "exclusive"],
     deps = [":train_lib"]
 )
@@ -38,8 +38,8 @@ py_test(
 py_test(
     name = "torch_quick_start",
     size = "small",
-    main = "examples/torch_quick_start.py",
-    srcs = ["examples/torch_quick_start.py"],
+    main = "examples/pytorch/torch_quick_start.py",
+    srcs = ["examples/pytorch/torch_quick_start.py"],
     tags = ["team:ml", "exclusive"],
     deps = [":train_lib"]
 )
@@ -71,19 +71,150 @@ py_test(
 py_test(
     name = "tune_cifar_torch_pbt_example",
     size = "medium",
-    main = "examples/tune_cifar_torch_pbt_example.py",
-    srcs = ["examples/tune_cifar_torch_pbt_example.py"],
+    main = "examples/pytorch/tune_cifar_torch_pbt_example.py",
+    srcs = ["examples/pytorch/tune_cifar_torch_pbt_example.py"],
     tags = ["team:ml", "exclusive", "pytorch", "tune"],
     deps = [":train_lib"],
     args = ["--smoke-test"]
 )
 
 py_test(
-    name = "tune_linear_example",
+    name = "tune_torch_regression_example",
     size = "small",
-    main = "examples/tune_linear_example.py",
-    srcs = ["examples/tune_linear_example.py"],
+    main = "examples/pytorch/tune_torch_regression_example.py",
+    srcs = ["examples/pytorch/tune_torch_regression_example.py"],
     tags = ["team:ml", "exclusive", "tune"],
+    deps = [":train_lib"],
+    args = ["--smoke-test"]
+)
+
+# Formerly AIR examples
+
+py_test(
+    name = "distributed_sage_example",
+    size = "small",
+    main = "examples/pytorch_geometric/distributed_sage_example.py",
+    srcs = ["examples/pytorch_geometric/distributed_sage_example.py"],
+    tags = ["team:ml", "exclusive", "gpu"],
+    deps = [":train_lib"],
+    args = ["--use-gpu", "--num-workers=2", "--epochs=1", "--dataset=fake"]
+)
+
+py_test(
+    name = "horovod_cifar_pbt_example",
+    size = "small",
+    srcs = ["examples/horovod/horovod_cifar_pbt_example.py"],
+    tags = ["team:ml", "exlusive"],
+    deps = [":train_lib"],
+    args = ["--smoke-test"]
+)
+
+py_test(
+    name = "horovod_pytorch_example",
+    size = "small",
+    srcs = ["examples/horovod/horovod_pytorch_example.py"],
+    tags = ["team:ml", "exclusive"],
+    deps = [":train_lib"],
+    args = ["--num-epochs=1"]
+)
+
+py_test(
+    name = "horovod_tune_example",
+    size = "small",
+    srcs = ["examples/horovod/horovod_tune_example.py"],
+    tags = ["team:ml", "exclusive"],
+    deps = [":train_lib"],
+    args = ["--smoke-test"]
+)
+
+py_test (
+    name = "huggingface_basic_language_modeling_example",
+    size = "medium",
+    srcs = ["examples/huggingface/huggingface_basic_language_modeling_example.py"],
+    args = ["--smoke-test", "--num-epochs 3"],
+    tags = ["team:ml", "exclusive"],
+    deps = [":train_lib"]
+)
+
+py_test(
+    name = "tensorflow_regression_example",
+    size = "medium",
+    main = "examples/tf/tensorflow_regression_example.py",
+    srcs = ["examples/tf/tensorflow_regression_example.py"],
+    tags = ["team:ml", "exclusive"],
+    deps = [":train_lib"],
+    args = ["--smoke-test"]
+)
+
+# This is tested in test_examples!
+# py_test(
+#     name = "tensorflow_mnist_example",
+#     size = "medium",
+#     main = "examples/tf/tensorflow_mnist_example.py",
+#     srcs = ["examples/tf/tensorflow_mnist_example.py"],
+#     tags = ["team:ml", "exclusive"],
+#     deps = [":train_lib"],
+#     args = ["--smoke-test"]
+# )
+
+# py_test(
+#     name = "tensorflow_autoencoder_example", # REGRESSION
+#     size = "medium",
+#     main = "examples/tf/tensorflow_autoencoder_example.py",
+#     srcs = ["examples/tf/tensorflow_autoencoder_example.py"],
+#     tags = ["team:ml", "exclusive"],
+#     deps = [":train_lib"],
+#     args = ["--smoke-test"]
+# )
+
+# This is tested in test_examples!
+# py_test(
+#     name = "torch_fashion_mnist_example",
+#     size = "medium",
+#     main = "examples/pytorch/torch_fashion_mnist_example.py",
+#     srcs = ["examples/pytorch/torch_fashion_mnist_example.py"],
+#     tags = ["team:ml", "exclusive"],
+#     deps = [":train_lib"],
+#     args = ["--smoke-test"]
+# )
+
+# py_test(
+#     name = "torch_fashion_mnist_example_gpu",
+#     size = "medium",
+#     main = "examples/pytorch/torch_fashion_mnist_example.py",
+#     srcs = ["examples/pytorch/torch_fashion_mnist_example.py"],
+#     tags = ["team:ml", "exclusive", "gpu"],
+#     deps = [":train_lib"],
+#     args = ["--use-gpu"]
+# )
+
+py_test(
+    name = "torch_regression_example",
+    size = "medium",
+    main = "examples/pytorch/torch_regression_example.py",
+    srcs = ["examples/pytorch/torch_regression_example.py"],
+    tags = ["team:ml", "exclusive"],
+    deps = [":train_lib"],
+    args = ["--smoke-test"]
+)
+
+# This is tested in test_examples!
+# py_test(
+#     name = "torch_linear_example",
+#     size = "small",
+#     main = "examples/pytorch/torch_linear_example.py",
+#     srcs = ["examples/pytorch/torch_linear_example.py"],
+#     tags = ["team:ml", "exclusive"],
+#     deps = [":train_lib"],
+#     args = ["--smoke-test"]
+# )
+
+py_test(
+    name = "tune_tensorflow_mnist_example",
+    size = "medium",
+    main = "examples/tf/tune_tensorflow_mnist_example.py",
+    srcs = ["examples/tf/tune_tensorflow_mnist_example.py"],
+    tags = ["team:ml", "exclusive"],
     deps = [":train_lib"],
     args = ["--smoke-test"]
 )
@@ -211,7 +342,7 @@ py_test(
     size = "large",
     srcs = ["tests/test_horovod_trainer.py"],
     tags = ["team:ml", "exclusive", "ray_air"],
-    deps = [":train_lib", "//python/ray/air:ml_lib"]
+    deps = [":train_lib", "//python/ray/air:train_lib"]
 )
 
 py_test(
