+++ python/ray/train/examples/mlflow_fashion_mnist_example.py
@@ -1,7 +1,7 @@
 import argparse
 
 from ray.air.config import RunConfig, ScalingConfig
-from ray.train.examples.torch_fashion_mnist_example import train_func
+from ray.train.examples.pytorch.torch_fashion_mnist_example import train_func
 from ray.train.torch import TorchTrainer
 from ray.air.callbacks.mlflow import MLflowLoggerCallback
