+++ python/ray/train/examples/pytorch/torch_fashion_mnist_example.py
