+++ python/ray/train/examples/tf/tensorflow_autoencoder_example.py
