+++ python/ray/train/examples/tf/tensorflow_mnist_example.py
