+++ python/ray/train/examples/tf/tune_tensorflow_mnist_example.py
